feat/create-repo-and-invite-teamMember

+++ src/TeamOverview.jsx
@@ -11,6 +11,8 @@ const TeamOverview = () => {
     const repoName = queryParams.get('repoName');
     const teamRepoId = queryParams.get('repoId');
     const teamName = queryParams.get('teamName');
+    const username = Cookies.get('username');
+    const token = Cookies.get('token');
     const teamId = queryParams.get('teamId');
 
     const [isModalOpen, setIsModalOpen] = useState(false);
@@ -80,6 +82,13 @@ const TeamOverview = () => {
     const deleteTeam = async () => {
 
         try {
+            const deleteGitResponse = await fetch(`http://localhost:3001/repo/delete?owner=${username}&repo=${repoName}&token=${token}`, {
+                method: 'POST'
+            });
+            if (!deleteGitResponse.ok) {
+                throw new Error('刪除git儲存庫時出錯');
+            }
+
             const deleteRepoResponse = await fetch(`http://localhost:8081/team-repos/${teamRepoId}`, {
                 method: 'DELETE'
             });
@@ -98,7 +107,7 @@ const TeamOverview = () => {
             }
 
             alert('成功刪除Github儲存庫');
-            
+
             navigate(`/teamRepo/?teamId=${teamId}`); // 重導向到首頁
             fetchTeamData();
         } catch (error) {
